refactor all commands into new pattern
